// src/api/types/ProjectEnvironment.ts
// This file was auto-generated by Fern from our API Definition.

/** The environment in which the server client is running */
export const ProjectEnvironment = {
    Development: "development",
    Production: "production",
} as const;
export type ProjectEnvironment = (typeof ProjectEnvironment)[keyof typeof ProjectEnvironment];